+++ app/src/main/java/kr/co/nottodo/data/remote/model/modification/ModificationDto.kt
@@ -1,30 +1,34 @@
-package kr.co.nottodo.data.remote.model
+package kr.co.nottodo.data.remote.model.modification
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestAdditionDto(
+data class RequestModificationDto(
     @SerialName("title") val title: String,
     @SerialName("situation") val situation: String,
     @SerialName("actions") val actions: List<String>?,
     @SerialName("goal") val goal: String?,
-    @SerialName("dates") val dates: List<String>,
 )
 
 @Serializable
-data class ResponseAdditionDto(
+data class ResponseModificationDto(
     @SerialName("status") val status: Int,
     @SerialName("success") val success: Boolean,
     @SerialName("message") val message: String,
-    @SerialName("data") val data: Addition,
+    @SerialName("data") val data: Modification,
 ) {
     @Serializable
-    data class Addition(
+    data class Modification(
         @SerialName("title") val title: String,
         @SerialName("situation") val situation: String,
-        @SerialName("actions") val actions: List<ResponseModificationDto.Modification.Action>?,
+        @SerialName("actions") val actions: List<Action>?,
         @SerialName("goal") val goal: String?,
-        @SerialName("dates") val dates: List<String>,
-    )
+    ) {
+        @Serializable
+        data class Action(
+            @SerialName("name")
+            val name: String,
+        )
+    }
 }
